// frontend/src/app/query/object-types.enum.ts
export enum ObjectTypesEnum {
    AS_BLOCK = "as-block",
    AS_SET = "as-set",
    AUT_NUM = "aut-num",
    DOMAIN = "domain",
    FILTER_SET = "filter-set",
    INET6NUM = "inet6num",
    INETNUM = "inetnum",
    INET_RTR = "inet-rtr",
    IRT = "irt",
    KEY_CERT = "key-cert",
    MNTNER = "mntner",
    ORGANISATION = "organisation",
    PEERING_SET = "peering-set",
    PERSON = "person",
    POEM = "poem",
    POETIC_FORM = "poetic-form",
    ROLE = "role",
    ROUTE = "route",
    ROUTE6 = "route6",
    ROUTE_SET = "route-set",
    RTR_SET = "rtr-set"
}